Rephrase LDAP check script output
The previous phrasing lead some people to believe that there is a limit
on the number of LDAP users that can sign in to a GitLab instance. That
is not the case; the limit in the check script only applies to the
diagnostic information result set, so that running `rake gitlab:check`
does not output thousands of LDAP users.